# backend/progresstracker/openapi.json
{
  "components" : {
    "schemas" : {
...
      "HabitRequest" : {
        "properties" : {
          "description" : {
            "maxLength" : 255,
            "minLength" : 0,
            "pattern" : "^[^\\x00-\\x1F\\x7F-\\x9F\\u2028\\u2029]*$",
            "type" : "string"
          },
          "frequency" : {
            "enum" : [
              "DAILY",
              "WEEKLY"
            ],
            "type" : "string"
          },
          "goalPeriod" : {
            "enum" : [
              "DAILY",
              "WEEKLY"
            ],
            "type" : "string"
          },
          "goalTargetCount" : {
            "format" : "int32",
            "maximum" : 7,
            "minimum" : 1,
            "type" : "integer"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 1,
            "pattern" : "^[^\\x00-\\x1F\\x7F-\\x9F\\u2028\\u2029]*$",
            "type" : "string"
          }
        },
        "required" : [
          "frequency",
          "name"
        ],
        "type" : "object"
      },
      "HabitResponse" : {
        "properties" : {
          "completedForPeriod" : {
            "description" : "Already completed for the current day (daily habits) or week (weekly habits), by the server's calendar, so a client never has to guess the date",
            "type" : "boolean"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currentStreak" : {
            "format" : "int32",
            "type" : "integer"
          },
          "description" : {
            "type" : "string"
          },
          "frequency" : {
            "enum" : [
              "DAILY",
              "WEEKLY"
            ],
            "type" : "string"
          },
          "goalPeriod" : {
            "enum" : [
              "DAILY",
              "WEEKLY"
            ],
            "type" : "string"
          },
          "goalTargetCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "lastCompletedDate" : {
            "format" : "date",
            "type" : "string"
          },
          "longestStreak" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "progressCount" : {
            "description" : "Completions so far in the current goal period",
            "format" : "int32",
            "type" : "integer"
          },
          "progressTargetCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "xpTotal" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [
          "completedForPeriod",
          "currentStreak",
          "frequency",
          "goalPeriod",
          "goalTargetCount",
          "id",
          "longestStreak",
          "name",
          "progressCount",
          "progressTargetCount",
          "xpTotal"
        ],
        "type" : "object"
      },
...
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "errors" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "description" : "Per-field validation messages. Present on validation failures only.",
            "type" : "object"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Habits, completions, streaks, XP, and achievements.",
    "title" : "Progress Tracker API",
    "version" : "1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/habits" : {
      "get" : {
        "operationId" : "getAllHabits",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/HabitResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid bearer token"
          }
        },
        "tags" : [
          "Habits"
        ]
      },
      "post" : {
        "operationId" : "createHabit",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/HabitRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HabitResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body is malformed or fails validation"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid bearer token"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The caller already has the most habits allowed"
          }
        },
        "tags" : [
          "Habits"
        ]
      }
    },
    "/api/habits/{id}" : {
      "delete" : {
        "operationId" : "delete",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "int64",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid bearer token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The habit belongs to another user"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "No such habit"
          }
        },
        "tags" : [
          "Habits"
        ]
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "int64",
              "type" : "integer"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/HabitRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HabitResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body is malformed or fails validation"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid bearer token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The habit belongs to another user"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "No such habit"
          }
        },
        "tags" : [
          "Habits"
        ]
      }
    },
    "/api/habits/{id}/complete" : {
      "post" : {
        "operationId" : "completeHabit",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "int64",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HabitResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid bearer token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The habit belongs to another user"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "No such habit"
          }
        },
        "tags" : [
          "Habits"
        ]
      }
    },
...
  },
  "security" : [
    {
      "bearerAuth" : [ ]
    }
  ]
}